==> notebooks/01-tratamento-padronizacao-dados.ipynb <==
# **Colab**

# === 1. Montagem do Google Drive e definição dos caminhos ===
from google.colab import drive
drive.mount('/content/drive')
caminho_base = '/content/drive/My Drive/Colab/cb-classificador/data/'
arquivo_entrada = caminho_base + 'item_nfe_rotulo.csv'
arquivo_tratado = caminho_base + 'dados_tratados.csv'

# === 2. Instalação e importação de bibliotecas ===
!pip install unidecode
from unidecode import unidecode
import pandas as pd
import csv
import nltk
nltk.download('stopwords')
from nltk.corpus import stopwords
stopwords_pt = set(stopwords.words('portuguese'))

# === 3. Função para limpeza e padronização dos campos ===
def limpar_campos_textuais(df):
    # Tratamento de valores nulos
    df['cod_anp'] = df['cod_anp'].fillna('0')
    df['cod_ncm'] = df['cod_ncm'].fillna('0')
    df['des_anp'] = df['des_anp'].fillna('nihil')
    df['des_anp'] = df['des_anp'].replace(['0', 'ANP'], 'nihil')
    df['des_produto'] = df['des_produto'].fillna('nihil')
    df['cod_unidade'] = df['cod_unidade'].fillna('desconhecida')
    # Conversão para inteiros
    df['cod_anp'] = df['cod_anp'].apply(lambda x: int(float(x)) if x != '' else 0)
    df['cod_ncm'] = df['cod_ncm'].apply(lambda x: int(float(x)) if x != '' else 0)
    # Limpeza e padronização de descrições
    for col in ['des_anp', 'des_produto']:
        df[col] = df[col].astype(str)
        df[col] = df[col].str.replace(r'\\', '', regex=True)
        df[col] = df[col].str.replace(',', '-')
        df[col] = df[col].str.lower().apply(unidecode)
    # Padronização de unidades
    df['cod_unidade'] = df['cod_unidade'].astype(str).str.lower().apply(unidecode)
    return df

# === 4. Função para padronizar abreviaturas de códigos de unidade ===
def padronizar_codigos_unidade(df, coluna='cod_unidade'):
    mapa_cod_unidade = {
        'u': 'un',
        'un1': 'un',
        'und': 'un',
        'uni': 'un',
        'unid': 'un',
        'unidad': 'un',
        'l': 'lt',
        'l.': 'lt',
        'lts': 'lt',
        'lit': 'lt',
        'litro': 'lt',
        'h': 'hr',
        'hs': 'hr',
        'pç': 'pc',
        'peca': 'pc'
    }
    df[coluna] = df[coluna].replace(mapa_cod_unidade)
    return df

# === 5. Função para substituir valores com baixa frequência ===
def substituir_valores_raros(df, colunas_limites, valor_substituto=9):
    for coluna, limite in colunas_limites.items():
        frequencias = df[coluna].value_counts()
        raros = frequencias[frequencias < limite].index
        df[coluna] = df[coluna].apply(lambda x: valor_substituto if x in raros else x)
    return df

# === 6. Função para remoção de stopwords dos campos textuais ===
def remover_stopwords(texto, stopwords_lista):
    if pd.isnull(texto):  # Verificar se o valor não é nulo
        return texto
    palavras = texto.split()
    palavras_filtradas = [palavra for palavra in palavras if palavra not in stopwords_lista]
    return ' '.join(palavras_filtradas)

# === 7. Função para remoção linhas duplicadas ===
def remover_duplicatas(df):
    return df.drop_duplicates()

# === 8. Função para aplicar o pipeline de tratamento e padronização dos dados ===
def tratar_dados(df, colunas_e_limites, stopwords_lista):
    df = limpar_campos_textuais(df)
    df = padronizar_codigos_unidade(df)
    df = substituir_valores_raros(df, colunas_e_limites)
    df['des_anp'] = df['des_anp'].apply(lambda x: remover_stopwords(x, stopwords_lista))
    df['des_produto'] = df['des_produto'].apply(lambda x: remover_stopwords(x, stopwords_lista))
    return df

# === 9. Leitura do arquivo de entrada ===
df = pd.read_csv(arquivo_entrada, skiprows=1, header=None, quotechar='"', delimiter=',')
df.columns = ['categoria', 'ano', 'num_chave_nota_fiscal_eletronica', 'num_item_nota_fiscal_eletronica', 'cod_anp', 'cod_ncm', 'cod_unidade', 'des_anp', 'des_produto']
df = df[['categoria', 'cod_anp', 'cod_ncm', 'cod_unidade', 'des_anp', 'des_produto']]

len(df)

df['categoria'].value_counts()

df['cod_ncm'].value_counts()

df['cod_anp'].value_counts()

df['cod_unidade'].value_counts()

# === 10. Aplicar o pipeline de tratamento e padronização dos dados ===
colunas_e_limites = {
    'cod_ncm': 10,
    'cod_anp': 10,
    'cod_unidade': 5
}
df1 = tratar_dados(df, colunas_e_limites, stopwords_pt)

df1['cod_ncm'].value_counts()

df1['cod_anp'].value_counts()

df1['cod_unidade'].value_counts()

# === 11. Remover duplicatas ===
df2 = remover_duplicatas(df1)

len(df2)

df2['categoria'].value_counts()

df2['cod_ncm'].value_counts()

df2['cod_anp'].value_counts()

df2['cod_unidade'].value_counts()

# === 14. Armazenar arquivo tratado ===
df2.to_csv(arquivo_tratado, index=False, quoting=csv.QUOTE_NONE, escapechar='\\')
