==> notebooks/04-treinamento-modelo-final.ipynb <==
# **Colab**

# === 1. Montagem do Google Drive e definição dos caminhos ===
from google.colab import drive
drive.mount('/content/drive')
arquivo_codificado = '/content/drive/My Drive/Colab/cb-classificador/data/dados_codificados.csv'
arquivo_modelo = '/content/drive/My Drive/Colab/cb-classificador/data/mlpclassifier_model.pkl'

# === 2. Importação de bibliotecas ===
import joblib
import pandas as pd
from sklearn.neural_network import MLPClassifier

# === 3. Leitura do arquivo de entrada ===
df_codificado = pd.read_csv(arquivo_codificado)

# === 4. Separação das variáveis preditoras (X) e alvo (y) ===
X = df_codificado.drop(columns=['categoria'])
y = df_codificado['categoria']

len(df_codificado)

len(df_codificado.columns)

df_codificado['categoria'].value_counts()

# === 5. Treinamento do modelo final com 100% dos dados ===
model = MLPClassifier(
    hidden_layer_sizes=(25,),
    activation='relu',
    solver='adam',
    alpha=0.001,
    max_iter=300,
    random_state=42
) # Instancia o algoritmo com a melhor configuração de hiperparâmetros indicada pelo GridSearchCV no notebook 03
model.fit(X, y)

# === 6. Armazenar modelo final ===
joblib.dump(model, arquivo_modelo) # Armazenar o modelo final para uso no protótipo
